+++ CHANGELOG.md
@@ -17,6 +17,9 @@
 * Launch Everyone in Random Directions:
   * Now gives all players a temporary feather to give you a chance at avoiding fall damage
 
+* Disable Equipment:
+  * Fixed Executive Card still being usable while effect is active
+
 * Increase & Decrease World Speed:
   * Fixed some player attack sounds playing with the wrong audio pitch